Made visualization archives renameable
Visualization archives, until now, had a top-level directory where all
files lived.  That's good practice for archives that are supposed to be
extracted at some point, but these aren't.  Removing this top-level
directory (and the assumption that its name matches the name of the
archive) makes it possible to rename the archives without editing them.

In other news, I changed the visinfo.cfg file to match the specification
in our Interim Project Report: the [graphs] section now specifies
filenames of the graphs, not titles.  This simplifies finding the graph
in the archive greatly.