Wrap the legend to fit on the display
In the netscience test file, there are so many colors that they can't
all fit in the width of 1920 pixels.  Amazing; also shows that we need
to seriously think about what to do when we run out of colors, because
we do on netscience.  Now we can at least see boxes for all the colors,
because they wrap onto more lines when the whole width of the window is
used up.  It might be good to make some way to hide it though, because
it really hogs screen space in the first and last stages where there are
over a hundred colors being added or removed.