User-specified color sets for decompositions
As per Felix's suggestion, now the user can select individual colors in
the decomposition visualization.  The decompositions on that color set
are automatically computed and displayed.  This can be slow; we should
certainly compute the new components and layout in a separate thread.
This seems like reasonable progress for tonight though.