Add a random button to Count, fix an infinite loop
I added a button to the toolbar in the Count stage which randomizes the
k-patterns and motifs highlighted.  This allows us to see a more
complete picture of what happened when counting motifs in this
component.

There was an infinite loop if the chosen vertex set had no "interesting"
k-patterns.  That's fixed now: we look at each k-pattern at most once,
and if none is interesting, we try a different vertex set instead of
hanging forever.