Indicate what vertices can go where in the top row
The vertices where the k-pattern boundary is mapped is black, the vertices
where the anonymous k-pattern vertices can be mapped are white, and the
vertices not in this k-pattern are the gray of the background.  This
will be followed by including a new row showing the k-patterns
themselves, alone, on the motif rather than the component.