Fix incorrect Count results after randomizing
The Count tab was showing incorrect results after clicking the "random"
button.  This was caused by a list not being cleared when we tried to
get a new random set of patterns, causing the display to be not updated
fully.  The problem has been fixed, so now everything makes sense and
life is good.